+++ src/main.js
@@ -29,6 +29,7 @@ export { default as HelpPopover } from './components/HelpPopover';
 export { default as ButtonIcon } from './components/ButtonIcon';
 export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
+export { default as Layout } from './components/Layout';
 export { default as Link } from './components/Link';
 export { default as ListGroup } from './components/ListGroup';
 export { default as OverlayWrapper } from './components/OverlayWrapper';